+++ ICD.Connect.Partitioning/Rooms/IRoomSettings.cs
@@ -28,6 +28,11 @@ namespace ICD.Connect.Partitioning.Rooms
 		Dictionary<int, eCombineMode> Sources { get; }
 
 		/// <summary>
+		/// Gets the audio destination settings.
+		/// </summary>
+		Dictionary<int, eCombineMode> AudioDestinations { get; }
+
+		/// <summary>
 		/// Gets the destination settings.
 		/// </summary>
 		Dictionary<int, eCombineMode> Destinations { get; }