+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+ - Audio destinations are now specifically listed as part of the room configuration
+
 ## [3.0.0] - 2018-04-11
 ### Added
  - VolumePoints are now Originators and are instantiated like devices, sources, etc